requirements.txt
streamlit
pytesseract
opencv-python
googletrans==4.0.0-rc1
Pillow